Close zero-send review accounting gaps
