+++ docs/DEVELOPMENT.md
@@ -1912,9 +1912,10 @@ Before every commit, verify the following:
   without a typed terminal HTTP status stay under the custody-lost/no-resend
   classification; with such a status they are retained as terminal actors for
   same-cycle replay, never as a second physical send. An identical envelope
-  never buys a settled actor twice, and a new retry cycle uses a new key. A `$0`
-  `not_dispatched` refusal remains retryable rather than becoming a sticky replay
-  actor, even though the host minted its synthetic operation id before admission.
+  never buys a settled actor twice, and a new retry cycle uses a new key. A
+  pre-write-ahead route/configuration/admission refusal is a retryable
+  `$0 not_dispatched` actor even though the host minted a synthetic operation id;
+  a later checkpoint failure cannot erase an already-fired paid stamp.
   Positive `settled`/`dispatched`/`unresolved` capture evidence outranks a
   contradictory synthetic `not_dispatched` label; only `reserved`/`released`
   is pre-dispatch. Reuse the physical-state vocabulary exported by
